Revert "fix: split test in attempt to evade CI badness"
This reverts commit e8055a475488feabb45597608252958235940010.
